+++ Xenity_Engine/Source/engine/tools/string_tag_finder.cpp
@@ -9,7 +9,7 @@
 bool StringTagFinder::FindTag(const std::string& textToSearchIn, const size_t index, const size_t textSize, const std::string& textToFind, int& startPosition, int& endPosition)
 {
 	bool found = false;
-	const int textToFindSize = textToFind.size();
+	const int textToFindSize = (int)textToFind.size();
 	bool notEquals = false;
 
 	for (int i = 0; i < textToFindSize; i++)
